fix(tests): separate functional and unit tests
