+++ tests/Keboola/S3ExtractorTest/Functional/OneFileFunctionalTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Keboola\S3ExtractorTest;
+namespace Keboola\S3ExtractorTest\Functional;
 
 use Keboola\S3Extractor\Extractor;
 use Monolog\Handler\TestHandler;
 use Monolog\Logger;
 
-class OneFileTest extends TestCase
+class OneFileFunctionalTest extends FunctionalTestCase
 {
     /**
      * @dataProvider initialForwardSlashProvider
@@ -32,7 +32,7 @@ class OneFileTest extends TestCase
 
         $expectedFile = $this->path . '/' . 'myfile.csv';
         $this->assertFileExists($expectedFile);
-        $this->assertFileEquals(__DIR__ . "/../../_data/file1.csv", $expectedFile);
+        $this->assertFileEquals(__DIR__ . "/../../../_data/file1.csv", $expectedFile);
         $this->assertTrue($testHandler->hasInfo("Downloading file /file1.csv"));
         $this->assertTrue($testHandler->hasInfo("Downloaded 1 file(s)"));
         $this->assertCount(2, $testHandler->getRecords());
@@ -62,7 +62,7 @@ class OneFileTest extends TestCase
 
         $expectedFile = $this->path . '/' . 'myfile.csv';
         $this->assertFileExists($expectedFile);
-        $this->assertFileEquals(__DIR__ . "/../../_data/folder1/file1.csv", $expectedFile);
+        $this->assertFileEquals(__DIR__ . "/../../../_data/folder1/file1.csv", $expectedFile);
         $this->assertTrue($testHandler->hasInfo("Downloading file /folder1/file1.csv"));
         $this->assertTrue($testHandler->hasInfo("Downloaded 1 file(s)"));
         $this->assertCount(2, $testHandler->getRecords());
